+++ src/drivers/generic/initrd/interface.h
@@ -14,11 +14,11 @@ namespace initrdInterface
 {
 	bool QueryFileProperties (
 		const char* path,
-		uint32_t driveId, uint8_t partitionIdOnDrive,
+		uint32_t driveId, uint32_t partitionIdOnDrive,
 		size_t* oFsizeBytes,
 		obos::driverInterface::fileAttributes* oFAttribs);
 	bool IterCreate(
-		uint32_t driveId, uint8_t partitionIdOnDrive,
+		uint32_t driveId, uint32_t partitionIdOnDrive,
 		uintptr_t* oIter);
 	bool IterNext(
 		uintptr_t iter,
@@ -28,7 +28,7 @@ namespace initrdInterface
 		obos::driverInterface::fileAttributes* oFAttribs);
 	bool IterClose(uintptr_t iter);
 	bool ReadFile(
-		uint32_t driveId, uint8_t partitionIdOnDrive,
+		uint32_t driveId, uint32_t partitionIdOnDrive,
 		const char* path,
 		size_t nToSkip,
 		size_t nToRead,
